final/requirements.txt
langchain_google_genai
langchain-community
grpcio
langchain
langchainhub
mitreattack-python 
pyyaml==6.0.1
fire==0.5.0
requests==2.28.1
attrs==21.2.0
pick==2.2.0
atomic-operator==0.9.0
langgraph
httpx
chromadb
attackcti
Jinja2
markdown
tiktoken
unstructured
nltk